requirements.txt
blinker==1.7.0
click==8.1.7
docutils==0.20.1
Flask==3.0.2
grpcio==1.62.1
grpcio-tools==1.62.1
itsdangerous==2.1.2
Jinja2==3.1.3
lockfile==0.12.2
MarkupSafe==2.1.5
pid==3.0.4
protobuf==4.25.3
python-daemon==3.0.1
service==0.6.0
setproctitle==1.3.3
setuptools==69.2.0
Werkzeug==3.0.1
